chore: movie store 세팅
