+++ frontend/movie-recommend/src/main.js
@@ -4,6 +4,11 @@ import { createPinia } from "pinia";
 import App from "./App.vue";
 import router from "./router";
 
+// Bootstrap
+import "bootstrap/dist/css/bootstrap.min.css";
+import "bootstrap-icons/font/bootstrap-icons.css";
+import "bootstrap/dist/js/bootstrap";
+
 const app = createApp(App);
 
 app.use(createPinia());
